stt-process/enhance/requirements.txt
annotated-types==0.7.0
anyio==4.6.0
boto3==1.35.29
botocore==1.35.29
certifi==2024.8.30
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
jiter==0.5.0
jmespath==1.0.1
openai==1.50.2
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3
python-dotenv==1.0.1
typing_extensions==4.12.2